requirements.txt
Django
django-dotenv
gunicorn
whitenoise
factory-boy
djangorestframework